+++ demo/start.sh
@@ -2,8 +2,8 @@
 
 rm -rf ./data/
 
-export BUILDX_HTTP_ADDR=0.0.0.0:6666
-export BUILDX_SSH_ADDR=0.0.0.0:6667
+export BUILDX_HTTP_ADDR=0.0.0.0:9910
+export BUILDX_SSH_ADDR=0.0.0.0:9911
 
 export BUILDX_INITIAL_USER=zhangjie
 export BUILDX_INITIAL_PASSWORD=zhangjie